requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==3.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.1
email_validator==2.2.0
etelemetry==0.3.1
fast-graphrag==0.0.5
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.11.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
igraph==0.11.8
importlib_resources==6.5.2
instructor==1.7.9
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
json_repair==0.39.1
looseversion==1.3.0
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
mysql-connector-python==8.0.26
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
openai==1.75.0
optype==0.9.3
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
prov==2.0.1
puremagic==1.28
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.1
PyJWT==2.1.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyMySQL==1.1.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.5
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9.1
ruff==0.11.5
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scipy-stubs==1.15.2.2
sentencepiece==0.2.0
shapely==2.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
texttable==1.7.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.15.2
tqdm==4.67.1
traits==7.0.2
transformers==4.36.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
vertexai==1.71.1
voyageai==0.3.2
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0